backend/pyproject.toml
[project]
name = "shrekify"
version = "0.1.0"
description = "Django API and React client for Shrekify image generation."
requires-python = ">=3.12"
dependencies = [
    "django==6.0",
    "djangorestframework==3.16.1",
    "django-cors-headers==4.9.0",
    "pillow==12.0.0",
    "diffusers>=0.30.0",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "accelerate>=0.26.0",
    "huggingface-hub>=0.26.0",
    "transformers>=4.45.0",
    "bitsandbytes>=0.48.2",
    "protobuf>=4.25.0",
    "sentencepiece>=0.2.0",
    "dotenv>=0.9.9",
    "ip-adapter>=0.1.0",
    "einops>=0.8.1",
    "xformers>=0.0.33.post2",
    "opencv-python>=4.12.0.88",
    "controlnet-aux>=0.0.10",
    "matplotlib>=3.10.8",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128"},
]
torchvision = [
  { index = "pytorch-cu128"},
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

